backend/plant_disease_recognition_model/requirements.txt
# Plant Disease Detection API - Requirements
# Python 3.11 or 3.12 recommended (avoid 3.13 for TensorFlow compatibility)

# Web Framework
fastapi[standard]>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# Machine Learning
tensorflow-macos>=2.16.0; sys_platform == 'darwin' and platform_machine == 'arm64'
tensorflow>=2.16.0; sys_platform != 'darwin' or platform_machine != 'arm64'
tf-keras>=2.16.0

# Image Processing
pillow>=10.0.0
numpy>=1.24.0,<2.0.0
